requirements.txt
h5py~=3.12.1
joblib==1.4.2
keras==3.8.0
matplotlib==3.9.4
numpy==1.23.5
openpyxl==3.1.5
pandas==2.2.3
protobuf==3.17.2
scikit-learn==1.6.1
sentencepiece==0.2.0
sympy==1.13.1
tensorflow==2.6.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.48.1
biopython~=1.85
tables==3.9.2